+++ astropy/utils/iers/tests/test_iers.py
@@ -13,6 +13,9 @@ from astropy import units as u
 from astropy.table import QTable
 from astropy.time import Time, TimeDelta
 
+
+TRAVIS = os.environ.get('TRAVIS', False)
+
 FILE_NOT_FOUND_ERROR = getattr(__builtins__, 'FileNotFoundError', OSError)
 
 try:
@@ -325,3 +328,38 @@ def test_IERS_B_parameters_loading_into_IERS_Auto():
             A[name][ok_A], B[name][i_B], rtol=1e-15,
             err_msg=("Bug #9206 IERS B parameter {} not copied over "
                      "correctly to IERS Auto".format(name)))
+
+
+# Issue with FTP, rework test into previous one when it's fixed
+@pytest.mark.xfail('TRAVIS')
+@pytest.mark.remote_data
+def test_iers_a_dl():
+    iersa_tab = iers.IERS_A.open(iers.IERS_A_URL, cache=False)
+    try:
+        # some basic checks to ensure the format makes sense
+        assert len(iersa_tab) > 0
+        assert 'UT1_UTC_A' in iersa_tab.colnames
+    finally:
+        iers.IERS_A.close()
+
+
+@pytest.mark.remote_data
+def test_iers_a_dl_mirror():
+    iersa_tab = iers.IERS_A.open(iers.IERS_A_URL_MIRROR, cache=False)
+    try:
+        # some basic checks to ensure the format makes sense
+        assert len(iersa_tab) > 0
+        assert 'UT1_UTC_A' in iersa_tab.colnames
+    finally:
+        iers.IERS_A.close()
+
+
+@pytest.mark.remote_data
+def test_iers_b_dl():
+    iersb_tab = iers.IERS_B.open(iers.IERS_B_URL, cache=False)
+    try:
+        # some basic checks to ensure the format makes sense
+        assert len(iersb_tab) > 0
+        assert 'UT1_UTC' in iersb_tab.colnames
+    finally:
+        iers.IERS_B.close()